Moved classes into own file
